Update field scroll box
