+++ packages/outside/src/components/SchoolClubFilter/index.tsx
@@ -1,28 +1,11 @@
 import { InputCancel } from '@bitgouel/common'
 import { useState } from 'react'
+import FieldScroll from '../FieldScroll'
 import * as S from './style'
-
-interface OptionProps {
-  value: string
-  label: string
-}
-
-const FilterOptions: OptionProps[] = [
-  { value: '미래형 운송기기', label: '미래형 운송기기' },
-  { value: '에너지산업', label: '에너지산업' },
-  { value: '의료•헬스케어', label: '의료•헬스케어' },
-  { value: 'AI 융•복합', label: 'AI 융•복합' },
-  { value: '문화산업', label: '문화산업' },
-]
 
 const SchoolClubFilter = () => {
   const [selectedOption, setSelectedOption] = useState<string>('핵심분야')
   const [isOpen, setIsOpen] = useState<boolean>(false)
-
-  const handleSelect = (option: OptionProps) => {
-    setSelectedOption(option.label)
-    setIsOpen(false)
-  }
 
   return (
     <S.ContSelect>
@@ -33,15 +16,10 @@ const SchoolClubFilter = () => {
         {selectedOption}
       </S.SelectContainer>
       {isOpen && (
-        <S.ListMember>
-          {FilterOptions.map((option) => (
-            <li key={option.value}>
-              <S.ListButton type='button' onClick={() => handleSelect(option)}>
-                {option.label}
-              </S.ListButton>
-            </li>
-          ))}
-        </S.ListMember>
+        <FieldScroll
+          setSelectedField={setSelectedOption}
+          setIsScrollOpen={setIsOpen}
+        />
       )}
 
       {selectedOption && selectedOption !== '핵심분야' && (
