+++ packages/outside/src/components/index.ts
@@ -8,3 +8,4 @@ export { default as SchoolClubFilter } from './SchoolClubFilter'
 export { default as SchoolClubInput } from './SchoolClubInput'
 export { default as SchoolContent } from './SchoolContent'
 export { default as SchoolItem } from './SchoolItem'
+export { default as FieldScroll } from './FieldScroll'
